model_service/requirements.txt
flask
torch
pickle5
Pillow
numpy
requests
scipy
Ninja
